+++ src/constants/wallet.config.ts
@@ -27,6 +27,7 @@ export const sessionParams = {
         'wallet_watchAsset',
         'wallet_switchEthereumChain',
         'wallet_addEthereumChain',
+        'eth_sendTransaction',
       ],
       chains: [`eip155:${CHAIN}`],
       events: ['chainChanged', 'accountsChanged'],
@@ -45,9 +46,23 @@ export const contractInfo = {
     devAddress: '0x9480c22ffdefb6d42923da54b731744229f61cfe' as Address,
     mainnetAddress: '0xda47Ba3A9F1DCb61C057Efe1e5d6b6654241c3Dd' as Address,
   },
+  USDT: {
+    symbol: 'USDT',
+    decimals: 18,
+    image: 'https://icons8.com/icon/DEDR1BLPBScO/tether',
+    devAddress: '0x337610d27c682E347C9cD60BD4b3b107C9d34dDd' as Address,
+    mainnetAddress: '0x55d398326f99059ff775485246999027b3197955' as Address,
+  },
+  USDC: {
+    symbol: 'USDC',
+    decimals: 18,
+    image: 'https://cryptologos.cc/logos/usd-coin-usdc-logo.png?v=029',
+    devAddress: '0x64544969ed7ebf5f083679233325356ebe738930' as Address,
+    mainnetAddress: '0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d' as Address,
+  },
 }
 
-export type ContractInfo = keyof typeof contractInfo
+export type ContractType = keyof typeof contractInfo
 export type ContractInfoReturns = {
   symbol: string
   decimals: number
@@ -55,7 +70,7 @@ export type ContractInfoReturns = {
   address: Address
 }
 
-export const getContractInfo = (type: ContractInfo): ContractInfoReturns => {
+export const getContractInfo = (type: ContractType): ContractInfoReturns => {
   const info = contractInfo[type]
   return {
     ...info,