+++ src/navigators/Navigators.tsx
@@ -16,10 +16,14 @@ import {isUserConfirmed} from 'utils/response'
 import ChangeEmail from 'screens/profile/changeEmail'
 import ResetPassword from 'screens/auth/ResetPassword'
 import RegistrationProgress from 'screens/auth/Registration'
+import CryptoPayment from 'screens/BuyToken/CryptoTransfer'
 import ChangePassword from 'screens/profile/changePassword'
 import BecomeAffiliate from 'screens/profile/becomeAffiliate'
 import RegistrationForm from 'screens/auth/Registration/RegisterForm'
+import OrderSummary from 'screens/BuyToken/BankTransfer/OrdersSummary'
+import DownLine from 'screens/affiliate/downLine/DownlinTable'
 import DeleteAccountConfirm from 'screens/profile/DeleteAccount/DeleteAccountConfirm'
+import PaymentInformation from 'screens/BuyToken/BankTransfer/PaymentInformation'
 
 import DrawerNavigator from './DrawerNavigator'
 import type {RouteStack} from './routes'
@@ -124,6 +128,38 @@ const Navigators = () => {
           component={DeleteAccountConfirm}
           options={{
             title: t('profile.deleteAccount.title'),
+            headerBackTitleVisible: false,
+          }}
+        />
+        <Stack.Screen
+          name='OrderSummary'
+          component={OrderSummary}
+          options={{
+            title: t('navigation.items.orderSummary'),
+            headerBackTitleVisible: false,
+          }}
+        />
+        <Stack.Screen
+          name='PaymentInformation'
+          component={PaymentInformation}
+          options={{
+            title: t('navigation.items.paymentInformation'),
+            headerBackVisible: false,
+          }}
+        />
+        <Stack.Screen
+          name='CryptoPayment'
+          component={CryptoPayment}
+          options={{
+            title: t('navigation.items.cryptoPayment'),
+            headerBackTitleVisible: false,
+          }}
+        />
+        <Stack.Screen
+          name='DownLine'
+          component={DownLine}
+          options={{
+            title: t('navigation.items.yourReferral'),
             headerBackTitleVisible: false,
           }}
         />