+++ CHANGELOG.md
@@ -1,6 +1,33 @@
 # Changelog
 
-## \[Unreleased v0.2.1\] - 2023-12-12
+## Platform APP v1.3.0 - 2024-01-17
+
+### Added
+
+- Added Bank Transfer
+- Added Bank Transfer PDF generate
+- Added user affiliate table tree views
+- Added bank details Pdf download form modal
+
+### Changed
+
+- Marimakko chat updated API
+- Hide Crypto payment for `USDT` and `USDC` Token
+- Checkbox hide in registrations when user come from Invite link
+
+### Removed
+
+- Referral Id from settings
+
+## Platform APP v1.2.0 - 2023-12-21
+
+### Added
+
+- Bank Transfer
+- Bank Transfer PDF generate
+- Crypto Payment for `USDT` and `USDC`
+
+## Platform APP v1.1.2 - 2023-12-12
 
 ### Added
 
@@ -14,7 +41,7 @@
 
 - Resolved issue with drop-down color for improved user experience
 
-## Platform APP v0.2.0 - 2023-11-28
+## Platform APP v1.1.0 - 2023-11-28
 
 ### Added